+++ app/scripts/fhir/sdc-common.js
@@ -84,23 +84,6 @@ function addCommonSDCFns(ns) {
 
 
   /**
-   * Find out if multiple answers extension is true.
-   * @param qItem - FHIR Questionnaire item.
-   * @returns {boolean}
-   */
-  self._hasMultipleAnswers = function (qItem) {
-    var ret = false;
-    if(qItem) {
-      var answerRepeats = LForms.Util.findObjectInArray(qItem.extension, 'url', self.fhirExtUrlAnswerRepeats);
-      if(answerRepeats && answerRepeats.valueBoolean) {
-        ret = true;
-      }
-    }
-    return ret;
-  };
-
-
-  /**
    * Do a shallow copy of specified fields from source to target.
    *
    * @param source - Source object