+++ test/data/STU3/cne-datatype-fixture.js
@@ -17,7 +17,6 @@ var STU3_cneDataTypeFixture = {
   },
 
   output: {
-    required: undefined,
     linkId: "/54126-8/54131-8",
     text: "Gender",
     type: "choice",
@@ -67,7 +66,6 @@ var STU3_alWithCodeSystemFixture = {
       {valueCoding: {code: "x", display: "X", system: "http://loinc.org"}},
       {valueCoding: {code: "y", display: "Y", system: "http://loinc.org"}},
       {valueCoding: {code: "z", display: "Z", system: "http://loinc.org"}}
-    ],
-    required: undefined
+    ]
   }
 };